Skip email step when email field has a value. (#5449)
