+++ changelog/task-1406-skip-email-step
@@ -1,0 +1,4 @@
+Significance: minor
+Type: add
+
+Skipping the email input step for WooPay express checkout flow when the email input field has already a value. The WooPay express checkout feature is behind a feature flag currently.
