+++ client/checkout/platform-checkout/email-input-iframe.js
@@ -7,35 +7,7 @@ import wcpayTracks from 'tracks';
 import request from '../utils/request';
 import showErrorCheckout from '../utils/show-error-checkout';
 import { buildAjaxURL } from '../../payment-request/utils';
-
-// Waits for the element to exist as in the Blocks checkout, sometimes the field is not immediately available.
-const waitForElement = ( selector ) => {
-	return new Promise( ( resolve ) => {
-		if ( document.querySelector( selector ) ) {
-			return resolve( document.querySelector( selector ) );
-		}
-
-		const checkoutBlock = document.querySelector(
-			'[data-block-name="woocommerce/checkout"]'
-		);
-
-		if ( ! checkoutBlock ) {
-			return resolve( null );
-		}
-
-		const observer = new MutationObserver( ( mutationList, obs ) => {
-			if ( document.querySelector( selector ) ) {
-				resolve( document.querySelector( selector ) );
-				obs.disconnect();
-			}
-		} );
-
-		observer.observe( checkoutBlock, {
-			childList: true,
-			subtree: true,
-		} );
-	} );
-};
+import { getTargetElement, validateEmail } from './utils';
 
 export const handlePlatformCheckoutEmailInput = async (
 	field,
@@ -44,7 +16,7 @@ export const handlePlatformCheckoutEmailInput = async (
 ) => {
 	let timer;
 	const waitTime = 500;
-	const platformCheckoutEmailInput = await waitForElement( field );
+	const platformCheckoutEmailInput = await getTargetElement( field );
 	let hasCheckedLoginSession = false;
 
 	// If we can't find the input, return.
@@ -303,7 +275,7 @@ export const handlePlatformCheckoutEmailInput = async (
 	} );
 
 	if ( isBlocksCheckout ) {
-		const formSubmitButton = await waitForElement(
+		const formSubmitButton = await getTargetElement(
 			'button.wc-block-components-checkout-place-order-button'
 		);
 		formSubmitButton.addEventListener( 'click', () => {
@@ -451,16 +423,6 @@ export const handlePlatformCheckoutEmailInput = async (
 			.finally( () => {
 				spinner.remove();
 			} );
-	};
-
-	const validateEmail = ( value ) => {
-		/* Borrowed from WooCommerce checkout.js with a slight tweak to add `{2,}` to the end and make the TLD at least 2 characters. */
-		/* eslint-disable */
-		const pattern = new RegExp(
-			/^([a-z\d!#$%&'*+\-\/=?^_`{|}~\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]+(\.[a-z\d!#$%&'*+\-\/=?^_`{|}~\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]+)*|"((([ \t]*\r\n)?[ \t]+)?([\x01-\x08\x0b\x0c\x0e-\x1f\x7f\x21\x23-\x5b\x5d-\x7e\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]|\\[\x01-\x09\x0b\x0c\x0d-\x7f\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]))*(([ \t]*\r\n)?[ \t]+)?")@(([a-z\d\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]|[a-z\d\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF][a-z\d\-._~\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]*[a-z\d\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF])\.)+([a-z\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]|[a-z\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF][a-z\d\-._~\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]*[0-9a-z\u00A0-\uD7FF\uF900-\uFDCF\uFDF0-\uFFEF]){2,}\.?$/i
-		);
-		/* eslint-enable */
-		return pattern.test( value );
 	};
 
 	const closeLoginSessionIframe = () => {
