+++ client/checkout/platform-checkout/express-button/express-checkout-iframe.js
@@ -3,9 +3,11 @@
  */
 import { __ } from '@wordpress/i18n';
 import { getConfig } from 'utils/checkout';
+import { getTargetElement, validateEmail } from '../utils';
 import wcpayTracks from 'tracks';
 
-export const expressCheckoutIframe = async ( api, context ) => {
+export const expressCheckoutIframe = async ( api, context, emailSelector ) => {
+	const platformCheckoutEmailInput = await getTargetElement( emailSelector );
 	let userEmail = '';
 
 	const parentDiv = document.body;
@@ -165,7 +167,7 @@ export const expressCheckoutIframe = async ( api, context ) => {
 		);
 		urlParams.append( 'wcpayVersion', getConfig( 'wcpayVersionNumber' ) );
 
-		if ( email ) {
+		if ( email && validateEmail( email ) ) {
 			urlParams.append( 'email', email );
 		}
 
@@ -250,5 +252,5 @@ export const expressCheckoutIframe = async ( api, context ) => {
 		}
 	} );
 
-	openIframe();
+	openIframe( platformCheckoutEmailInput?.value );
 };
