+++ client/checkout/platform-checkout/express-button/index.js
@@ -37,6 +37,7 @@ const renderPlatformCheckoutExpressButton = () => {
 						'data-product_page'
 					)
 				}
+				emailSelector="#billing_email"
 			/>,
 			platformCheckoutContainer
 		);
